cDAC DBI: normalize async ParseContinuation diagnosticIP to CORDB_ADDRESS
Per review feedback (dotnet/runtime#130275): rather than shrinking the managed
cDAC out-param to a host-sized nuint, normalize the DBI IPC value to a fixed
64-bit CORDB_ADDRESS on the native side so a 64-bit DBI can consume a 32-bit
target (cross-bitness), matching how VMPTR/CORDB_ADDRESS fields are always 8 bytes.

ICorDebugProcess12::GetAsyncStack -> CordbAsyncStackWalk calls
IDacDbiInterface::ParseContinuation, whose native ABI declared the out-param as
PCODE* pDiagnosticIP. PCODE == TADDR is host/target-pointer-sized (4 bytes on x86,
8 on x64), and the RS caller passed the address of a 4-byte PCODE stack local. The
managed cDAC modeled the parameter as ulong* and wrote 8 bytes through it,
overrunning the x86 stack slot and tripping the /GS cookie
(STATUS_STACK_BUFFER_OVERRUN, 0xC0000409); it passed on x64 where PCODE is 8 bytes.

Fix by widening the value to a fixed 64-bit CORDB_ADDRESS end to end:
 - dacdbiinterface.h / dacdbiimpl.h / dacdbiimpl.cpp: PCODE* -> CORDB_ADDRESS*
   (the DAC widens PCODE to CORDB_ADDRESS on write).
 - rsstackwalk.cpp (CordbAsyncStackWalk::PopulateFrame/Next) and process.cpp
   (GetAsyncStack): local PCODE diagnosticIP -> CORDB_ADDRESS (downstream
   CordbAsyncFrame already used CORDB_ADDRESS).
 - Managed IDacDbiInterface.cs / DacDbiImpl.cs: keep the out-param as ulong*
   (fixed 64-bit), matching the widened native ABI.

Validated on Windows x86 Debug: all 7 previously-crashing Async tests pass
(AsyncBreakpoint, AsyncBreakpointJmc, AsyncGeneric, AsyncRecursive,
AsyncSharedGeneric, AsyncSimple, AsyncV2CallingAsyncV1Iterator).

+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Legacy/Dbi/IDacDbiInterface.cs
@@ -816,7 +816,7 @@ public unsafe partial interface IDacDbiInterface
     int GetAssemblyFromModule(ulong vmModule, ulong* pVmAssembly);
 
     [PreserveSig]
-    int ParseContinuation(ulong continuationAddress, nuint* pDiagnosticIP, ulong* pNextContinuation, uint* pState);
+    int ParseContinuation(ulong continuationAddress, ulong* pDiagnosticIP, ulong* pNextContinuation, uint* pState);
 
     [PreserveSig]
     int EnumerateAsyncLocals(ulong vmMethod, ulong codeAddr, uint state,
